requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.4.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-unfold==0.72.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Flask==3.1.2
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.7.1
mysqlclient==2.2.7
orjson==3.11.2
packaging==25.0
phonenumbers==9.0.2
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pygraphviz==1.14
PyJWT==2.10.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
twilio==9.10.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.5.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
